Optimize 2020 day 25
The original version, was very slow.

Since all values are always less than MODULUS, simple multiplication
followed by taking the modulo is safe. This approach could fail for a
MODULUS > 4294967296, but the MODULUS is hard-coded and smaller than
that, and the simple multiplication is dramatically faster.

Finding cycles by keeping a hashset was incredibly slow. Again, since
MODULUS is small, and values are all smaller than MODULUS, by the
pigeon-hole principle if we check all the values up to MODULUS loops
we'll have hit all values we'll see.

With these two fixes in place, the code runs about 30x faster with
runtimes on my machine dropping from ~6s to roughly ~0.2s.