Update 2020 day 8 to use an efficient algorithm
This approach is a little more complicated than the brute force
approach, but should perform the fix in O(n) time on the input size.

On my system the old code ran in ~130ms, and the new code closer to
100ms, so the difference is pretty minimal, but for larger inputs the
difference would get large, and I'm honestly surprised that this was any
improvement at all!