Linked list solution for day 20
This is absolutely not the most efficient solution. Some approach
tracking indices and using vectors is almost certainly more efficient,
but I'm kind of proud of myself for making a working linked list
solution in Rust!

I'm going to have to scrap this if I want to get the runtime down though
- currently it needs about 350ms which is more than I'm willing to
  budget for day 20!