2021 day 21 in Rust
This was a fun one.

I figure out the number of games played/won each round for each player
individually, and then combine them to get the total win counts.

In retrospect, a dynamic programming approach would probably have been
easier to come up with, and would have solved the problem plenty fast,
but this solution runs in 0.5ms on my computer, and could scale up to
much higher scores.