2021 day 22 in Rust
This solution is kind of slow - full runtime on my computer is a little
under 2 minutes. But the code is easy to reason about and clean.
Somewhat shockingly, I wrote this code from scratch, found a single bug
in the `intersection` function, and then it passed all test cases and
parts with no futzing!

So I may some day come back to this problem and try for a more efficient
algorithm. I'm confident that my `add` method for the `CuboidSet` is
inefficient algorithmically. But for now, I'm happy to move on!