2023 day 7
I went a little nuts with the generics here, but I'm pretty pleased that
I avoided duplicating logic, or implementing any explicit ordering logic
anywhere at all!

The code is a little long, but a lot of the length is just enum
definitions and match statements, which seems ok to me.

At <1ms runtime I don't even mind that this isn't the most efficient
possible implementation.