2023 day 11
I'm pretty pleased with the efficiency on this day.

Initial runtime was around 6ms, and I managed to get it down to 1.1ms by
switching to prefix sums for calculating the number of voids. Prefix
sums can be calculated once on initialization instead of ~100k times
when calculating all the distances.