2023 day 17
Simple Dijkstra's algorithm solution.

The run time is a little higher than I'd like, but it's a clean enough
solution for a first go. I'm going to probably need to substantially
improve runtime on both today and day 16 if I want to stay on track for
the 1s goal though!

Optimizations to consider:
- Tweak `State` to be easier to hash.
- Swap out BinaryHeap
    - Maybe for something like https://github.com/mpdn/radix-heap
- Switch to A* algorithm
    - A heuristic just measuring distance from the end should be
      admissible, and will encourage search down and to the right first
      which might be a huge win.
- For the ultra crucibles, skip over the sections where we can't turn
    - We can skip 3 blocks after each turn safely