2023 day 18
I'm pretty happy with this solution. Directly calculating areas is nice.

I would like to go back at some point and add logic to validate that the loop
is in fact a simple polygon (no self-intersection), but given that
that's clearly built into the problem setup, it isn't a high priority.