2023 day 192023 day 19
This was a little involved, but the runtime is good, and the solution is
clean enough.

Checking for intersections at the end takes a full 0.5ms, and clearly
isn't necessary for the actual inputs, but I think it's still correct to
do the check. Even if it's pretty clear the problem was designed to work
this way, there's nothing in the problem statement that guarantees this
behavior.