Readability tweaks to 2023 day 19
This is just a few tweaks for readability. The only functional change is
to stop pre-calculating the output ranges since apparently that has no
measurable effect on performance. Calculating output ranges directly is
easiert to reason about!

Other than that, this commit just has some minor re-shuffling and
variable renaming to make things a little clearer.