2023 day 21
This one got away from me a little, but I'm happy with the balance of
performance and generality.

The code is a little long, but the core logic is very simple and most
of the code is various checks that would only be needed to prevent
errors when processing inputs other than the official ones.