2023 day 22
I'm pretty happy with the runtime on this one.

I went with a brute force O(n^2) algorithm for part 2, and that seems to
be more than fast enough. I think something like an O(n log(n))
algorithm probably exists, but it would likely be a lot more
complicated, and at 1.7ms total runtime, and most of that for parsing
and part1, there's just not a ton of time to be gained here.