2023 day 23
The code is a little long, but I'm pretty pleased with the performance.

Basically all the run time is in `longest_path`, and I don't think any
non-brute force approach can work here. Since the inner loop is doing
some pretty minimal and quick operations, I'm happy with the approach
even if it is taking 50% longer than all the other problems combined!