2023 day 24
Nice clean z3 based solution.

There's definitely room for performance improvement here. Even using z3
I could cut the runtime down to ~70ms by just using the first three
hailstones. The problem with this approach is that it relies on the
first three hailstones having enough information (which isn't logically
guaranteed).

I also suspect that a purely algebraic solution is possible which would
dramatically improve runtime, so I'm not really inclined to try to speed
up the z3 solution at the cost of lost correctness or generality.

I may come back to this problem and try for the algebraic solution
later, but for now I'm still well under my 1s goal for the year, and
want to press ahead and finish all the problems.