2023 Day 25
I solved this one with Karger's algorithm. The solution is reasonably
simple and pleasant, but probabilistic and way too slow.

I'm going to rewrite this to hit my performance goals.

I could almost certainly improve performance by using Karger-Stein and
tightening up my implementation (in particular to avoid the hashmap).
That said, looking around at other people's solutions, it looks as if
Karger's algorithm is just slow for this problem. I'm going to switch to
Stoer-Wagner, or possibly some simpler probabilistic algorithm. There
are some very fast probabilistic approaches out there, though I'll need
to find one I'm confident is only probabilistic in runtime, not in
outcome if I'm going to do one of those.

I'm a little surprised by that since Karger's aglorithm was presented as
a fast algorithm that sacrifices determinism for speed and is efficient
for unweighted graphs (that's why I chose it in the first place), but
maybe that was a misunderstanding on my part.