Faster algorithm for 2023 day 25
This approach is hugely faster. Pick two points and drop paths between
them until 3 times. If we have two components, then we've found our
component sizes! If we still have a single component try another point
for our second point.

This algorithm doesn't guarantee that we find a unique cut or that the
cut is in fact a min-cut, but if we a 3-cut exists this algorithm will
find it, and the problem statement essentially promises that the cut
exists and is unique.

If the two components are of roughly equal size (as seems to be the case
in the real inputs), we will generally find the components after 1-2
iterations with exponentially decreasing probability of longer runtimes.
So while the worst case scenario runtime is potentially slow, the
average runtime on real inputs is going to be very fast and quite
consistent.