Speed up 2023 day 24
The algebraic solution is much faster!

With this, I think I'm done with my immediate goals for 2023. While
there are still optimizations to be had, most of the runtime is now in
day 23, and I don't think there's much chance of speeding that up other
than by parallelization.

Maybe if I'm feeling compulsive I'll see if I can strip a few
tens of miliseconds off some of the slower problems, but on the whole,
I'm pretty happy with where this landed!