2015 day 13 in Rust
My initial solution checking all permutations was a little cleaner and
looked a lot like the Python solution. However, this approach duplicated
a lot of effort since for a round table the permutations included each
ordering n times (where n is the number of people).

Fixing the first person at the table gives a factor of n speed increase
for the slow part of the program, and seems pretty worthwhile even if
the code is less pretty, so I decided to go with it.

This could probably be sped up further by using something like Heap's
algorithm to do the permutations in place on an array of indices. I
haven't looked closely at the `Itertools` approach, but since it's
designed to work on iterators and returns owned vectors it seems likely
to be a lot slower than an in-place permutation generation algorithm.

That said, I'm not currently trying to write highly optimized code for
2015, and my benchmarks are showing something like 2.2ms total runtime
so I'm not inclined to do that right now!