Use Heap's algorithm for 2015 day 13
So I got compulsive and went ahead and implemented this. The speedup is
actually substantial (proportionally) - runtime went from 2.2ms to
1.1ms so this more than doubled the speed of the core loop.

I would like to have implemented `Iterator` for this, but that can't be
done soundly since if we give out a reference to the current permutation
and then ask for the next permutation the value will change.

Given that I couldn't implement `Iterator` I'm fairly pleased with the
interface!