2015 day 17 in Rust
As before, fixing bugs in the python solutions as I go.

It's probably possible to do an iterator version of partitions, which
would obviously be necessary to handle large inputs. That said, for the
numbers of permutations involved here (~1300), I'm pretty sure the
iterator version would not only be more complex, but would actually
perform worse.