2015 day 19 in Rust
I went kind of overboard on the parsing.

I thought about trying to add logic validate the rules would work for
cheat-algorithm I'm using, but ultimately the only way to guarantee that
the substitution we want exists would be to actually find the
substitutions, and I couldn't bring myself to give up the simple
arithmetic solution.