2015 day 20 in Rust
I didn't like the the runtime of the naive algorithm in Python, but
really liked the nice simple code. After some fiddling I eventually
settled on searching for the upper bound by doing ever increasing
calculations which provides something like a 3-5x speed up over the
naive approach for very little added complexity.

It's very satisfying that I can get that sort of speed up even starting
with the kind of ridiculous starting limit of 1.

Half the total runtime comes from the last run, and a quarter from the
second to last, and so on. So any choice of starting limit that isn't
quite close to correct is almost as likely to increase runtime as
decrease it. Starting at 400k does worse than starting at 1 because we
do runs at 800k and then 1.6M. while the last two runs starting from 1
are at ~500k, and ~1M. Since I couldn't come up with a principled reason
to start at a higher value, 1 is fine!