2015 day 24 in Rust
This finishes up 2015 in Rust!

I went a little nuts with this one with the custom bitset and the
memoization on the subset sum function. A simpler solution that does
neither can still run in just a few seconds.

Performance is quite good though at ~35ms total runtime. I'm fairly
pleased with how well encapsulated the lower level code is. Most of the
lines of code come down to `PackageSet` implementation and tests and it's
pretty easy to read without having to look at either of those.