Begin updating 2019 Rust code
I'm planning to rework pretty much all of this code to match the style
elsewhere in the repo before finishing up the last few problems.

This change involves a big grab bag of changes that shouldn't have
changed functionality, but got everything running with modern Rust and
with clippy happy.

I still intend to:
- rework error handling to use `anyhow` in the binaries,
- rework the library code as feels appropriate, and
- go through each problem for style and obvious fixes.