Speed up 2023 day 23
Take advantage of the fact that the graph is planar to prune it before
trying to calculate the longest path.

The fact that the graph is planar means that whenever you've walked from
the entrance to the an edge of the map one of the possible directions
can't connect to the exit without crossing the path. I use DFS to figure
out if a given node is on the edge, and then don't add an edge for the
useless direction.

The current approach assumes the entrance and exit are in the expected
places, and disables the optimization if they're not. In principle it
should be possible to handle other entrance/exit placements, but I
didn't feel like implementing the more general logic right now when the
simple version works.

Total runtime for 2023 is now sub 150ms, which is pretty cool!