2023 day 17 performance tweak
It turned out that a big chunk of runtime was just going to vector
allocations. Using `Vec::with_capacity()` eliminates basically all of
that saving a surprising 8.6ms.

I could switch to an iterator and avoid allocations altogether, but I
think that the compiler is smart enough to allocate the vector just once and
then re-use that same memory for each call. In that case, the added
overhead of the `Iterator` approach would likely just increase runtime.

At this point about 35% of the total runtime is in `BinaryHeap::pop`
with nothing else making up a big slice. I could probably boost
performance by writing a faster priority queue that's less general than
`std::collections::BinaryHeap`, but I'm not really interested in doing
that.

Other substantial performance gains would have to come from reducing the
number of iterations. There may be ways to prune the graph, or a well
thought out A* implementation might be faster.