Rework 2023 day 16, also update runtimes
This shaves a chunk of time off by precalculating the sets of tiles
illuminated for each segment of the grid starting and ending at either
the map edge or a splitter.

This is a good start, but there are substantial improvements still
available if I can cache not only the segments but the whole
`EnergizedSet`s lit up by a single splitter.