2024 day 2
The solution for part 2 is O(n^2), and an O(n) solution is absolutely
possible. That said, I've already had a go, and the O(n) solution has
surprisingly tricky edge cases (well, maybe not that tricky, but all the
obvious solutions are so ugly!). With total runtime of 1.8ms, I'm not
super motivated to make uglier code right now, so I'll probably come
back to this later, but it's fine for now!