2024 day 4
I've had yet another go at tweaking my planar library. I'm still not
totally satisfied with it, but I think it's improving in usability.

Performance is so-so for this one, but I'm not going to trim a full
milliscond of the runtime, and big-O performance is as good as it's
going to get, so I'm not inclined to fuss about it.