Start optimizing day 6
Instead of retracing the whole path, start right in front of the
obstruction added for part 2. This dramatically speeds up runtime.

There's still room for optimization here. I'm pretty confident that
with better data structures and eliminating unecessary/redundant
calculation I can get this below 50ms, and probably in the 30ms
neighborhood, but for now, it's time to start day 7.