Optimize 2024 day 7
Going through the numbers in reverse order and using the inverse to
check if a value could produce the right result basically does tree
prunning on the search, leading to dramatically reduced runtime.

For multiply and concatenate in particular, the inverse is rarely going
to exist for a random pairing of numbers, so this prevents us from
needing the large majority of checks.

Over all I'm very happy with this solution - it's clean, readable, and
very efficient!