Further optimize 2024 day 6
We can't just start right in front of the new obstruction carelessly
because if we've passed through a point before, it's possible adding the
obstruction kicks us off the map on the first pass stopping us from ever
getting to the second pass. But if we ensure that we only test each
point once, that problem goes away, and we're left with a very efficient
approach!

I also finally got around to adding checks for loops in the initial run.
As far as I can tell the inputs are all built so that that can't happen,
but it's still better to verify we're going to terminate!