2024 day 12
This solution is a little over complicated and messy, and definitely
inefficent, but it works well enough.

On reflection, I can make this both simpler and faster by counting
corners instead of my messy and complicated side merging, but I'm going
to commit the working solution before I set to optimizing!