Optimize 2024 day 12
Switching to finding corners from messy side-merging saves a lot of
runtime!

I'm not thrilled with my code on this one. It feels a little messy and
confused, so I may go back and clean it up. But for the time being it
hits my main goals of correct and fast, so I'm happy to move on!