2024 day 14
I'm fairly pleased with how this worked out. The solution is fast, and
basically correct. I could, in principle, rewrite `has_picture` to do
something smarter and make it completely correct, but given that a
simple `return true` would also accomplish the purpose, I'm not sure I
feel very strongly about it.

In any case, I'm really happy that I've got what ammounts to a at least
the sketch of a fully general and correct solution that still runs very
quickly on the real input.