2024 day 17
This solution isn't perfect. An efficient solution pretty much requires
reverse engineering the program. This solution should work for any
program where each output only depends on the last 3 bits of the 'a'
register. Ideally I would validate that the program provided does that,
but for now this will do!

It might also be possible to make this a little faster by caching VM
outputs, but it's already fast enough that I'm not sure there's much
real gain to be had, and honestly, running the VM for a single output is
pretty quick.